new artwork

+++ app/src/main/java/org/xvdr/recordings/model/PacketAdapter.java
@@ -13,7 +13,7 @@ public class PacketAdapter {
         Movie movie = new Movie();
 
         movie.setTimeStamp(p.getU32() * 1000L);
-        movie.setDuration(p.getU32());
+        movie.setDuration((int)p.getU32());
         p.getU32(); // Priority
         p.getU32(); // Lifetime
         movie.setChannelName(p.getString()); // ChannelName
